requirements.txt
PyMuPDF==1.26.1
anthropic==0.54.0
camelot-py==1.0.0
faiss-cpu==1.11.0
langchain-anthropic==0.3.15
langchain==0.3.25
langgraph==0.4.8
langchain-community==0.3.25
sentence-transformers==2.3.1
langchain-openai==0.3.24
numpy==1.26.2
pandas==2.3.0
matplotlib==3.8.2
seaborn==0.13.2
python-dotenv==1.0.0
pydantic==2.10.4
typing-extensions==4.12.2
